ai-sre/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-sre"
version = "0.1.0"
description = "AI SRE Agent System — Claude Agent SDK + MCP orchestration"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.56",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "httpx>=0.28.0",
    "clickhouse-connect>=0.8.0",
    "pyyaml>=6.0",
    "prometheus-client>=0.22.0",
    "structlog>=24.4.0",
    "pydantic>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8.0",
    "mypy>=1.14",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
